requirements.txt
coverage==4.3.4
dj-database-url==0.4.2
Django==1.11
django-countries==4.4
django-crispy-forms==1.6.1
django-filter==1.0.2
django-model-utils==2.6.1
djangorestframework==3.6.2
hiredis==0.2.0
psutil==5.2.2
psycopg2==2.7.1
pytz==2017.2
redis==2.10.5
redislite==3.2.311
rules==1.2
six==1.10.0
tblib==1.3.2
